ref(anthropic)!: Move Anthropic AI integration to `@sentry/server-utils`
Move the Anthropic gen-ai instrumentation from `@sentry/core` to
`@sentry/server-utils/src/ai/anthropic-ai`. `instrumentAnthropicAiClient` is
re-exported by node/cloudflare/vercel-edge from `@sentry/server-utils`, so the
platform SDK public API is unchanged.

Also makes `SpanStatusType` public on `@sentry/core` (the Anthropic integration
is the only AI integration that consumes it).
