+++ packages/cloudflare/src/index.ts
@@ -77,7 +77,6 @@ export {
   // eslint-disable-next-line typescript/no-deprecated
   inboundFiltersIntegration,
   instrumentGoogleGenAIClient,
-  instrumentAnthropicAiClient,
   eventFiltersIntegration,
   linkedErrorsIntegration,
   requestDataIntegration,
@@ -127,7 +126,7 @@ export { httpServerIntegration } from './integrations/httpServer';
 export { fetchIntegration } from './integrations/fetch';
 export { spotlightIntegration } from './integrations/spotlight';
 export { vercelAIIntegration } from './integrations/tracing/vercelai';
-export { prismaIntegration, instrumentOpenAiClient } from '@sentry/server-utils';
+export { prismaIntegration, instrumentOpenAiClient, instrumentAnthropicAiClient } from '@sentry/server-utils';
 
 // eslint-disable-next-line typescript/no-deprecated
 export { instrumentD1WithSentry } from './instrumentations/worker/instrumentD1';
