+++ packages/node/src/index.ts
@@ -30,7 +30,7 @@ export {
   vercelAiIntegration as vercelAIIntegration,
 } from '@sentry/server-utils/orchestrion';
 export { redisIntegration } from './integrations/tracing/redis';
-export { prismaIntegration, instrumentOpenAiClient } from '@sentry/server-utils';
+export { prismaIntegration, instrumentOpenAiClient, instrumentAnthropicAiClient } from '@sentry/server-utils';
 export { setupHapiErrorHandler } from './integrations/tracing/hapi';
 export { setupKoaErrorHandler } from './integrations/tracing/koa';
 export {
@@ -133,7 +133,6 @@ export {
   updateSpanName,
   supabaseIntegration,
   instrumentSupabaseClient,
-  instrumentAnthropicAiClient,
   instrumentGoogleGenAIClient,
   zodErrorsIntegration,
   profiler,
