+++ packages/server-utils/src/ai/index.ts
@@ -5,3 +5,4 @@
  */
 
 export { instrumentOpenAiClient } from './openai';
+export { instrumentAnthropicAiClient } from './anthropic-ai';
